Removes old email delivery code
