+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/components/ResponseDelivery.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { useCallback, useState, useMemo, useEffect } from "react";
+import React, { useCallback, useState, useEffect } from "react";
 import {
   LocalizedFormProperties,
   FormServerError,
@@ -7,12 +7,9 @@ import {
 } from "@lib/types/form-builder-types";
 import { useTranslation } from "@i18n/client";
 import { useSession } from "next-auth/react";
-import { isValidGovEmail } from "@lib/validation/validation";
-import { ResponseEmail } from "@formBuilder/components/ResponseEmail";
 import { Radio } from "@formBuilder/components/shared/MultipleChoice";
 import { Button } from "@clientComponents/globals";
 import { useTemplateStore } from "@lib/store/useTemplateStore";
-import { completeEmailAddressRegex } from "@lib/utils/form-builder";
 import { FormPurposeHelpButton } from "./dialogs/FormPurposeHelpButton";
 import { ResponseDeliveryHelpButtonWithApi } from "./dialogs/ResponseDeliveryHelpDialogApiWithApi";
 import {
@@ -21,7 +18,6 @@ import {
 } from "@formBuilder/components/ClassificationSelect";
 import {
   sendResponsesToVault,
-  updateTemplateDeliveryOption,
   updateTemplateSecurityAttribute,
   updateTemplateFormPurpose,
 } from "@formBuilder/actions";
@@ -40,8 +36,8 @@ import { EventKeys, useCustomEvent } from "@lib/hooks/useCustomEvent";
 
 enum DeliveryOption {
   vault = "vault",
-  email = "email",
   api = "api",
+  email = "email", // Deprecated
 }
 
 /*
@@ -55,11 +51,9 @@ export enum PurposeOption {
   nonAdmin = "nonAdmin",
 }
 
-// TODO: replace with ResponseDeliveryWithoutEmail once emailDeliverly feature flag is removed
-export const ResponseDelivery = () => {
+export const ResponseDeliveryWithoutEmail = () => {
   const { t, i18n } = useTranslation("form-builder");
   const { status } = useSession();
-  const session = useSession();
   const { refreshData } = useRefresh();
   const lang = i18n.language === "en" ? "en" : "fr";
   const { apiKeyId } = useFormBuilderConfig();
@@ -70,10 +64,6 @@ export const ResponseDelivery = () => {
     id,
     resetDeliveryOption,
     updateField,
-    subjectEn: initialSubjectEn,
-    subjectFr: initialSubjectFr,
-    defaultSubjectEn,
-    defaultSubjectFr,
     securityAttribute,
     updateSecurityAttribute,
     isPublished,
@@ -107,7 +97,7 @@ export const ResponseDelivery = () => {
     t("formSettingsModal.emailOption.label")
   );
 
-  const userEmail = session.data?.user.email ?? "";
+  // Used to check for legacy forms that have email delivery option
   let initialDeliveryOption = !email ? DeliveryOption.vault : DeliveryOption.email;
 
   const hasApiKey = apiKeyId ? true : false;
@@ -117,62 +107,10 @@ export const ResponseDelivery = () => {
     initialDeliveryOption = DeliveryOption.api;
   }
 
-  const [deliveryOptionValue, setDeliveryOptionValue] = useState(initialDeliveryOption);
+  const [deliveryOptionValue, setDeliveryOptionValue] = useState(
+    initialDeliveryOption as DeliveryOption
+  );
   const [purposeOption, setPurposeOption] = useState(formPurpose as PurposeOption);
-
-  const [inputEmailValue, setInputEmailValue] = useState(email ? email : userEmail);
-  const [subjectEnValue, setSubjectEnValue] = useState(
-    initialSubjectEn ? initialSubjectEn : defaultSubjectEn
-  );
-  const [subjectFrValue, setSubjectFrValue] = useState(
-    initialSubjectFr ? initialSubjectFr : defaultSubjectFr
-  );
-
-  const [isInvalidEmailError, setIsInvalidEmailError] = useState(false);
-
-  /*--------------------------------------------*
-   * Form Validation
-   *--------------------------------------------*/
-  const isValid = useMemo(() => {
-    if (classification !== securityAttribute) {
-      return true;
-    }
-
-    const isValidDeliveryOption =
-      !isInvalidEmailError &&
-      inputEmailValue !== "" &&
-      subjectEnValue !== "" &&
-      subjectFrValue !== "";
-    const emailDeliveryOptionsChanged =
-      inputEmailValue !== email ||
-      subjectEnValue !== initialSubjectEn ||
-      subjectFrValue !== initialSubjectFr;
-
-    if (deliveryOptionValue === DeliveryOption.email) {
-      if (!completeEmailAddressRegex.test(inputEmailValue)) {
-        return false;
-      }
-      return isValidDeliveryOption && emailDeliveryOptionsChanged;
-    }
-
-    if (deliveryOptionValue === initialDeliveryOption) {
-      return false;
-    }
-
-    return true;
-  }, [
-    deliveryOptionValue,
-    initialDeliveryOption,
-    isInvalidEmailError,
-    inputEmailValue,
-    email,
-    subjectEnValue,
-    initialSubjectEn,
-    subjectFrValue,
-    initialSubjectFr,
-    classification,
-    securityAttribute,
-  ]);
 
   /*--------------------------------------------*
    * Set as Database Storage
@@ -183,81 +121,29 @@ export const ResponseDelivery = () => {
     });
 
     if (!result.error) {
-      // Update local state
-      setInputEmailValue("");
-
       // Update the template store
       resetDeliveryOption();
       updateSecurityAttribute(classification);
     }
 
     return result;
-  }, [id, resetDeliveryOption, setInputEmailValue, classification, updateSecurityAttribute]);
-
-  /*--------------------------------------------*
-   * Set as Email Delivery
-   *--------------------------------------------*/
-  const setToEmailDelivery = useCallback(async () => {
-    if (!isValidGovEmail(inputEmailValue)) return false;
-
-    // Call server action
-    const result = await updateTemplateDeliveryOption({
-      id,
-      deliveryOption: {
-        emailAddress: inputEmailValue,
-        emailSubjectEn: subjectEnValue,
-        emailSubjectFr: subjectFrValue,
-      },
-    });
-
-    if (!result.error) {
-      // Update the template store
-      updateField("deliveryOption.emailAddress", inputEmailValue);
-      updateField("deliveryOption.emailSubjectEn", subjectEnValue);
-      updateField("deliveryOption.emailSubjectFr", subjectFrValue);
-      updateSecurityAttribute(classification);
-    }
-
-    return result;
-  }, [
-    inputEmailValue,
-    subjectEnValue,
-    subjectFrValue,
-    id,
-    updateField,
-    classification,
-    updateSecurityAttribute,
-  ]);
+  }, [id, resetDeliveryOption, classification, updateSecurityAttribute]);
 
   /*--------------------------------------------*
    * Save Delivery Option
    *--------------------------------------------*/
   const saveDeliveryOptions = useCallback(async () => {
-    let result;
-
-    if (
-      (email !== "" && deliveryOptionValue === DeliveryOption.vault) ||
-      deliveryOptionValue === DeliveryOption.api
-    ) {
-      // Call local callBack which will call the server action
-      result = (await setToDatabaseDelivery()) as FormServerError;
-    } else {
-      // Call local callBack which will call the server action
-      result = (await setToEmailDelivery()) as FormServerError;
-    }
-
-    if (result?.error) {
+    const resultDelivery = (await setToDatabaseDelivery()) as FormServerError;
+    if (resultDelivery?.error) {
       toast.error(<ErrorSaving errorCode={FormServerErrorCodes.DELIVERY_OPTION} />, "wide");
       return;
     }
 
-    // Update security attribute server action
-    result = (await updateTemplateSecurityAttribute({
+    const resultAttribute = (await updateTemplateSecurityAttribute({
       id,
       securityAttribute: classification,
     })) as FormServerError;
-
-    if (result?.error) {
+    if (resultAttribute?.error) {
       toast.error(<ErrorSaving errorCode={FormServerErrorCodes.DELIVERY_OPTION} />, "wide");
       return;
     }
@@ -265,16 +151,7 @@ export const ResponseDelivery = () => {
     toast.success(t("settingsResponseDelivery.savedSuccessMessage"));
 
     refreshData && refreshData();
-  }, [
-    t,
-    refreshData,
-    email,
-    deliveryOptionValue,
-    id,
-    classification,
-    setToDatabaseDelivery,
-    setToEmailDelivery,
-  ]);
+  }, [t, refreshData, id, classification, setToDatabaseDelivery]);
 
   /*--------------------------------------------*
    * Save form purpose option
@@ -365,34 +242,28 @@ export const ResponseDelivery = () => {
                 </p>
               ) : null}
 
-              {/* Email Option */}
-              {!hasApiKey && (
-                <div className="mb-10">
-                  <Radio
-                    disabled={isPublished || protectedBSelected || hasApiKey}
-                    id={`delivery-option-${DeliveryOption.email}`}
-                    checked={deliveryOptionValue === DeliveryOption.email}
-                    name="response-delivery"
-                    value={DeliveryOption.email}
-                    label={emailLabel}
-                    onChange={updateDeliveryOption}
-                    className="mb-0"
-                  />
-                  {deliveryOptionValue === DeliveryOption.email && (
-                    <ResponseEmail
-                      inputEmail={inputEmailValue}
-                      setInputEmail={setInputEmailValue}
-                      subjectEn={subjectEnValue}
-                      setSubjectEn={setSubjectEnValue}
-                      subjectFr={subjectFrValue}
-                      setSubjectFr={setSubjectFrValue}
-                      isInvalidEmailError={isInvalidEmailError}
-                      setIsInvalidEmailError={setIsInvalidEmailError}
+              {/* Email Option - only show for published legacy forms that have this value */}
+              {isPublished && deliveryOptionValue === DeliveryOption.email && (
+                <>
+                  <div className="mb-8 w-3/5 rounded-md bg-indigo-50 p-3">
+                    <p className="mb-2 font-bold">
+                      {t("settingsResponseDelivery.emailDeliveryDeprecated.title")}
+                    </p>
+                    <p>{t("settingsResponseDelivery.emailDeliveryDeprecated.description")}</p>
+                  </div>
+                  <div className="mb-10">
+                    <Radio
+                      disabled={true}
+                      id={`delivery-option-${DeliveryOption.email}`}
+                      checked={true}
+                      name="response-delivery"
+                      value={DeliveryOption.email}
+                      label={emailLabel}
+                      className="mb-0"
                     />
-                  )}
-                </div>
+                  </div>
+                </>
               )}
-              {/* End Email Option */}
 
               {/* Vault Option */}
               {!hasApiKey && (
@@ -462,7 +333,8 @@ export const ResponseDelivery = () => {
               {deliveryOptionValue !== DeliveryOption.api && !hasApiKey && (
                 <>
                   <Button
-                    disabled={!isValid || isPublished}
+                    // disabled={!isValid || isPublished}
+                    disabled={isPublished}
                     theme="secondary"
                     onClick={saveDeliveryOptions}
                   >
