+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/page.tsx
@@ -2,7 +2,7 @@ import { serverTranslation } from "@i18n";
 import { Metadata } from "next";
 import { ApiKeyDialog } from "../components/dialogs/ApiKeyDialog/ApiKeyDialog";
 import { DeleteApiKeyDialog } from "../components/dialogs/DeleteApiKeyDialog/DeleteApiKeyDialog";
-import { ResponseDeliveryWithoutEmail } from "./components/ResponseDeliveryWithoutEmail";
+import { ResponseDeliveryWithoutEmail } from "./components/ResponseDelivery";
 
 export async function generateMetadata(props: {
   params: Promise<{ locale: string }>;
